requirements.txt
Flask
torch
torchvision
opencv-python
Pillow
flask-cors
numpy